+++ src/Symfony/Component/Serializer/Normalizer/DateTimeNormalizer.php
@@ -25,17 +25,17 @@ final class DateTimeNormalizer implements NormalizerInterface, DenormalizerInter
     public const FORMAT_KEY = 'datetime_format';
     public const TIMEZONE_KEY = 'datetime_timezone';
     public const CAST_KEY = 'datetime_cast';
+    public const SUPPORTED_TYPES_KEY = 'supported_types';
 
     private array $defaultContext = [
         self::FORMAT_KEY => \DateTimeInterface::RFC3339,
         self::TIMEZONE_KEY => null,
         self::CAST_KEY => null,
-    ];
-
-    private const SUPPORTED_TYPES = [
-        \DateTimeInterface::class => true,
-        \DateTimeImmutable::class => true,
-        \DateTime::class => true,
+        self::SUPPORTED_TYPES_KEY => [
+            \DateTimeInterface::class => true,
+            \DateTimeImmutable::class => true,
+            \DateTime::class => true,
+        ],
     ];
 
     public function __construct(array $defaultContext = [])
@@ -50,7 +50,7 @@ final class DateTimeNormalizer implements NormalizerInterface, DenormalizerInter
 
     public function getSupportedTypes(?string $format): array
     {
-        return self::SUPPORTED_TYPES;
+        return $this->defaultContext[self::SUPPORTED_TYPES_KEY];
     }
 
     /**
@@ -138,13 +138,7 @@ final class DateTimeNormalizer implements NormalizerInterface, DenormalizerInter
 
     public function supportsDenormalization(mixed $data, string $type, ?string $format = null, array $context = []): bool
     {
-        foreach (self::SUPPORTED_TYPES as $supportedType => $supported) {
-            if (!$supported) {
-                continue;
-            }
-            return $type === $supportedType || is_subclass_of($type, $supportedType);
-        }
-        return isset(self::SUPPORTED_TYPES[$type]);
+        return isset($this->defaultContext[self::SUPPORTED_TYPES_KEY][$type]);
     }
 
     /**